tucanbot/requirements.txt
requests==2.22.0
lxml==4.4.1
python-telegram-bot==12.0.0
